+++ crates/djangofmt_lint/src/checker.rs
@@ -52,7 +52,7 @@ impl<'a> Checker<'a> {
     /// Returns whether the given rule should be checked.
     #[must_use]
     #[inline]
-    pub fn is_rule_enabled(&self, rule: Rule) -> bool {
+    pub const fn is_rule_enabled(&self, rule: Rule) -> bool {
         self.settings.is_enabled(rule)
     }